ENG-01: extract _common (DataMatrix, epsqrt, _nz, SpecificationWarning)
First leaf module of the _pca_pls split. _pca_pls re-imports these names
so the public surface and the _nz test import stay intact.
